refactor: Use dependency inversion to avoid props drilling
